+++ core/apis_test.go
@@ -14,7 +14,7 @@ import (
 
 //Services
 
-func TestGetVersion(t *testing.T) {
+func TestSerGetVersion(t *testing.T) {
 	storage := genmocks.Storage{}
 	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil)
 	got := coreAPIs.GetVersion()