+++ utils/utils.go
@@ -1,54 +1,14 @@
 package utils
 
 import (
-	crand "crypto/rand"
 	"encoding/base64"
-	"log"
 	"math/rand"
-	"net/http"
-	"time"
 )
-
-//LogRequest logs the request as hide some header fields because of security reasons
-func LogRequest(req *http.Request) {
-	if req == nil {
-		return
-	}
-
-	method := req.Method
-	path := req.URL.Path
-
-	header := make(map[string][]string)
-	for key, value := range req.Header {
-		var logValue []string
-		//do not log sensitive information
-		if key == "Authorization" || key == "Csrf" {
-			logValue = append(logValue, "---")
-		} else {
-			logValue = value
-		}
-		header[key] = logValue
-	}
-	log.Printf("%s %s %s", method, path, header)
-}
-
-//RandSeq generates a random alphanumeric string of n characters
-func RandSeq(n int) string {
-	var letters = []rune("abcdefghijklmnopqrstuvwxyz0123456789")
-
-	rand.Seed(time.Now().UnixNano())
-
-	b := make([]rune, n)
-	for i := range b {
-		b[i] = letters[rand.Intn(len(letters))]
-	}
-	return string(b)
-}
 
 // GenerateRandomBytes returns securely generated random bytes
 func GenerateRandomBytes(n int) ([]byte, error) {
 	b := make([]byte, n)
-	_, err := crand.Read(b)
+	_, err := rand.Read(b)
 	if err != nil {
 		return nil, err
 	}